[coro] Separated that into impl/coro.hpp
